+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTConstructorDeclaration.java
@@ -65,7 +65,7 @@ public class ASTConstructorDeclaration extends AbstractMethodOrConstructorDeclar
      * (excluding any receiver parameter). A varargs parameter counts as one.
      */
     public int getArity() {
-        return getFormalParameters().getParameterCount();
+        return getFormalParameters().size();
     }
 
     //@Override // enable this with PMD 7.0.0 - see interface ASTMethodOrConstructorDeclaration