+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTMethodDeclarator.java
@@ -31,7 +31,7 @@ public class ASTMethodDeclarator extends AbstractJavaNode {
      */
     @Deprecated
     public int getParameterCount() {
-        return getFirstChildOfType(ASTFormalParameters.class).getParameterCount();
+        return getFirstChildOfType(ASTFormalParameters.class).size();
     }
 
     /**