+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symboltable/MethodNameDeclaration.java
@@ -23,12 +23,12 @@ public class MethodNameDeclaration extends AbstractNameDeclaration {
 
     public boolean isVarargs() {
         ASTFormalParameters params = (ASTFormalParameters) node.getChild(0);
-        if (params.getParameterCount() == 0) {
+        if (params.size() == 0) {
             return false;
         }
 
         // If it's a varargs, it HAS to be the last parameter
-        ASTFormalParameter p = (ASTFormalParameter) params.getChild(params.getParameterCount() - 1);
+        ASTFormalParameter p = (ASTFormalParameter) params.getChild(params.size() - 1);
         return p.isVarargs();
     }